Don't try to resume non-job-controlled job
The `continue_job' function requires a job-controlled job, but the
`fg_builtin' function was not checking whether the selected job was job-
controlled in case no operands were given to the built-in. As a result,
the function was trying to resume the non-job-controlled job or aborting
because of an assertion error.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3559 048f04df-13f5-43d7-8114-9f9ceecaec24