Handle out-of-range history entries correctly
Assertions in `fc_builtin' were wrong. If `vfirst' is a non-existing
entry number beyond the last entry, `lfirst.next' is `Histlist'.
Likewise, if `vlast' is a non-existing entry number before the first,
`llast.prev' is `Histlist'. In such cases, the first or last entry must
be chosen as a fallback.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3561 048f04df-13f5-43d7-8114-9f9ceecaec24