Parse command syntax "fc -s foo=bar n" correctly
When the "fc" built-in was used with the "-s" option, the first operand
of the form "old=new" was being parsed after the number of the operands
were tested for syntax error. As a result, the second operand was being
treated as an error.

This commit moves part of code so that the first operand is parsed
before the number of the operands are tested.

Note that it is not a correct fix to modify the test for the number of
the operands because it would incorrectly accept the syntax with two
operands where the first is not of the form "old=new".

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3563 048f04df-13f5-43d7-8114-9f9ceecaec24