Don't forget job on automatic job report (#35551)
POSIX requires the job-controlling interactive shell to report the exit
status of finished background jobs before each prompt. However, the exit
status must be available in a next invocation of the "wait" built-in.
Thus, the exit status must not be forgotten when reporting.

POSIX allows, when the -b option is also on, the shell to forget the
exit status. For brevity of implementation, yash now remembers the exit
status in this case.

https://osdn.jp/ticket/browse.php?group_id=3863&tid=35551

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3570 048f04df-13f5-43d7-8114-9f9ceecaec24