Forget awaited jobs
The previous commit has modified the shell so that jobs are reported but
not removed from the job list in some situations. As a side effect, the
"wait" built-in was no longer removing finished jobs that had already
been reported because the `print_job_status' function was just ignoring
finished jobs.

This commit so modifies the `print_job_status' function that finished
jobs are removed regardless of whether they are being reported.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3571 048f04df-13f5-43d7-8114-9f9ceecaec24