Suppress trivial error from test
In test-p.tst, the "find" command scans the "/dev" directory to find a
block and character special file that can be used in the test. On some
environments, the command prints an error message like "/dev/fd/3 is not
found". We can just ignore the error because the command is only
required to find a single special file rather than scanning the entire
directory thoroughly.


git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3572 048f04df-13f5-43d7-8114-9f9ceecaec24