Remove quotes after empty field removal
This is a refactoring commit that adds a quote removal step in the word
expansion process. As required by POSIX, quote removal is performed
after field splitting.

Unlike POSIX, quote removal is performed after empty field removal
because presence of quotation characters affects empty field removal.

Still unlike POSIX, quote removal is performed before pathname expansion
(globbing) because the current implementation of pathname expansion does
not support single- and double-quotation characters. This does not break
POSIX conformance since quoted characters are correctly escaped before
they are passed to the pathname expansion step.

Currently, the quote removal step added in this commit is a no-op
because quotes are already removed in the four expansions (in the
expand_word_inner function). The upcoming commit will modify this
behavior so that the new step is actually effective.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/branches/emptyfieldremoval@3588 048f04df-13f5-43d7-8114-9f9ceecaec24