Don't remove quotes too early (#35866)
Previously, single- and double-quotation marks were removed in the
expand_word_inner function. It was a cause of the bug where the
expansion of ""$space"" was resulting in no word while two empty words
are expected.

After this commit, quotation marks remain in the result of the function.
The marks are now removed as expected in the quote removal step, which
was added in the previous commit.

The "putempty" flag in the expand_word_T structure is replaced with the
"zeroword" flag, which is set only when "$@" is expanded for an empty
list of positional parameters. Callers of the expand_word_inner function
examine the flag and remove the unwanted empty word.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/branches/emptyfieldremoval@3589 048f04df-13f5-43d7-8114-9f9ceecaec24