Merge expand_string to expand_single
As a result of the previous commit, the expand_string and expand_single
functions are quite similar. This refactoring commit removes
expand_string in favor of expand_single. Two parameters are added to
expand_single so that it can behave like expand_string.

The expand_single_and_unescape auxiliary function is added.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3604 048f04df-13f5-43d7-8114-9f9ceecaec24