Escape backslash in $IFS when joining words
The concatenate_values function was treating any first character of $IFS
as a normal character. When called from the expand_single function, the
result of concatenation is subject to unescaping but the $IFS character
was not being properly escaped. As a result, if the $IFS character was a
backslash, it was misinterpreted as an escape.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3605 048f04df-13f5-43d7-8114-9f9ceecaec24