Don't remove empty last field in field splitting (#35283)
The POSIX specification for field splitting is intended to match the
behavior of Korn shell, which does not produce an empty field as the
last field if the last IFS non-whitespace is not followed by any non-IFS
characters. This commit modifies yash's default behavior to match Korn
shell's.

The new shell option 'empty-last-field' is added. By enabling this
option, users can restore the previous behavior.

Korn shell still fails the "fsplit-p.tst" tests because of its behavior
in empty field removal and quote removal. This issue is not addressed in
this commit.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3608 048f04df-13f5-43d7-8114-9f9ceecaec24