read: Don't leave IFS whitespaces in results
When the read built-in reads more fields than given variables, the
built-in is required to concatenate excessive fields into one.
Previously, if the first of the concatenated fields is empty and
delimited by IFS whitespaces followed by a IFS non-whitespace, the
built-in was leaving the IFS whitespaces, as well as the IFS
non-whitespace. Although the POSIX specification is vague about whether
this is a compliant behavior, this is counterintuitive and does not
match other shells' behavior.

After this commit, the built-in no longer leaves the IFS whitespaces in
the resultant variable value.

The actual modification in this commit is in the field splitting
implementation, since inclusion of IFS whitespaces is due to the results
of field splitting. The extract_fields function was previously returning
pointers to the beginning of the IFS whitespaces as the beginning and
end of the empty field. This was causing the IFS whitespaces to be
included in the concatenated variable value. After this commit, the
function returns pointers to the IFS non-whitespace that just follows
the IFS whitespaces, excluding them from the concatenation.

Note that the return value of the extract_fields function is not
modified in this commit.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3609 048f04df-13f5-43d7-8114-9f9ceecaec24