Reject unclosed here-documents in non-POSIX mode
Previously, when the POSIXly correct mode was off, the shell was
silently accepting unclosed here-documents. This behavior allowed users
to omit the end-of-here-document indicator at the end of a script.
However, this might also leave a misspelled indicator unnoticed, causing
everything up to the end of script to be treated as here-document
content rather than shell commands.

It would be better to require an explicit end-of-here-document indicator
for each here-document so that the user can ensure the here-document is
closed as intended.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3620 048f04df-13f5-43d7-8114-9f9ceecaec24