Apply errexit to expansion error in case (#35304)
This commit modifies the behavior of the case compound command so that
the errexit option is applied to expansion errors that occur in the case
command.

While behavior of existing shells disagree on this edge case, the author
believes the errexit option should generally applied to all kinds of
errors.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3629 048f04df-13f5-43d7-8114-9f9ceecaec24