Let SIGINT interrupt line-editing (#36327)
POSIX requires line-editing to be interruptible by the SIGINT signal.
This commit implements that requirement.

Note that the default key binding for Ctrl-C in line-editing had already
been interruption. Since Ctrl-C during line-editing does not produce
SIGINT and is directly handled by the shell, this commit does not affect
typical use of Ctrl-C. This commit is rather for handling SIGINT that
could be sent from another process.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3633 048f04df-13f5-43d7-8114-9f9ceecaec24