Suppress uninitialized variable warnings safely
Previously, the INIT macro was doing a self-assignment to deceive the
compiler that falsely claims the declared variable is used without
initialization. However, the current GCC is so smart that it treats a
self-assignment as an undefined behavior, which results in too
aggressive optimization that makes the program behave unexpectedly.

We still want to suppress the false warnings, but should avoid self-
assignments in doing so.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3673 048f04df-13f5-43d7-8114-9f9ceecaec24