Predict command successions more aggressively
This commit makes two changes in the probability computation algorithm
for command line prediction.

First, the candidate probability for the first previous command line is
halved. Previously, the previous command had twice as high probability
as the second previous, which proved too distinctive. Now the first
previous command has a probability of 1/2, the second 1/3, the third
1/4, and so on.

Second, the probability of candidate command lines that share preceding
commands with the current context is increased. Now there is a separate
descending probability sequence for each count of matching preceding
commands up to 4. This should make prediction results reflect more
command succession patterns.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3714 048f04df-13f5-43d7-8114-9f9ceecaec24