Refactor execstate_T
Previously, there were two types of members in the execstate_T
structure. One was set by a loop executor to affect behavior of specific
built-ins. The other was set by the built-ins to affect behavior of loop
executors. For understandability, this refactoring commit moves the
latter out of the structure.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3718 048f04df-13f5-43d7-8114-9f9ceecaec24