Reset execstate before changing lexical scope (#36747, #36766)
This commit modifies the function executor, variable executor, and the
dot built-in so that the function execution state is reset by
`reset_execstate' before a function, variable, or dot script is
executed. As a result, the "break" and "continue" built-ins are no
longer able to jump out of them.

(#36747) The POSIX specification has been loosened about the behavior of
the "break" and "continue" built-ins. They are now only required to
interrupt a lexically enclosing loop of them. After this commit, they
will fail if they have no lexically enclosing loop to interrupt, even
when there is a non-lexically enclosing loop some other shell might
allow to interrupt.

(#36766) This commit also fixes a bug where the "return" built-in fails
to return from a dot script when the shell is interactive.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3723 048f04df-13f5-43d7-8114-9f9ceecaec24