Make assignments on function temporary (#36880)
The effect of assignments in a simple command was clarified in
POSIX.1-2008 TC2 (2016). The existing behavior of yash already meets the
new requirements, but they allow different behaviors.

Until this commit, variables assigned in a simple command that calls a
function remained after the function returns. This behavior was
inconsistent with variables assigned in a simple command that executes
an external command, which are only effective during the command. To
make functions behave more like external commands, variables should not
remain after functions return.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3752 048f04df-13f5-43d7-8114-9f9ceecaec24