Initialize $YASH_LOADPATH only if missing
The $YASH_LOADPATH variable is now initialized on shell startup only if
not set from the environment.

This is required because the shell is going to be so changed that it
loads a default yashrc from the load path if ~/.yashrc is not found. In
most automated test cases, however, the shell should load neither
~/.yashrc or the default yashrc. The test runner bypasses this trap by
setting $YASH_LOADPATH to an empty string.


git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3756 048f04df-13f5-43d7-8114-9f9ceecaec24