Rewrite wait built-in
This refactoring commit extracts the big body of a loop out of the wait
built-in, making a new function.

As a side effect, the exit status of the wait built-in is different
between before and after the refactoring if the last operand is an
invalid job specification and a job specified by any other operand exits
with a non-zero status. Before the refactoring, the built-in returned
the exit status of the job, but it now returns Exit_FAILURE.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3764 048f04df-13f5-43d7-8114-9f9ceecaec24