Return POSIXly-correct exit status in wait built-in
According to POSIX, the "wait" built-in should return the exit status of
the last specified job only if all the job specifications are valid. If
any job specification is invalid, the built-in should return an exit
status between 1 and 126.

This commit modifies computation of the exit status so that the built-in
never returns an exit status of 127 if any job specification is invalid.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3765 048f04df-13f5-43d7-8114-9f9ceecaec24