Remove unstable test
The test case 'describing regular built-in with -b option' was dependent
on the "array" regular built-in, which is not always compiled in. This
commit removes the test case.

We could skip the test case rather than removing it, but that would not
be very meaningful. The pre-check to see if the test case should be
skipped would be done by the "command -vb" built-in and the test case
itself is also the "command -vb" built-in.

Note that there is no regular built-in that is always compiled in
regardless of configuration.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3768 048f04df-13f5-43d7-8114-9f9ceecaec24