Complete longer abbreviated Git commit ref
Abbreviated commit refs are typically 7 digits long, but they can be
longer in a large repository. Candidates should not be filtered on the
basis of their length.

At most ten candidates are now generated so that the user is not
overwhelmed by a mass of candidates (which may take a long time to be
generated).

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3776 048f04df-13f5-43d7-8114-9f9ceecaec24