Don't count finished jobs in \j in $PS1 (#37099)
As a result of r3570 and r3571, finished jobs are not automatically
removed from the job list when their status change is reported. This was
causing the job count printed by the \j notation in prompt strings ($PS1
and $PS2) to include those finished jobs until they are waited for by
the "wait" or "fg" built-ins.

This behavior was rather confusing because many users think those jobs
are removed once they are reported. Such jobs should not be included in
the job count.

Bug reported by Leah Neukirchen

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3780 048f04df-13f5-43d7-8114-9f9ceecaec24