Print job status update if SIGCHLD is not received
The "handle_sigchld" function was calling the "print_job_status_all"
function if a SIGCHLD had been received. However, "print_job_status_all"
needs to be called if there is any updated job, regardless of whether a
SIGCHLD has been received.

The '\j in PS1 and -b option' test in prompt-y.tst was failing if the
echo command was implemented as an external command rather than a
built-in. The external echo command makes yash receive a SIGCHLD, which
flags "print_status" in "handle_sigchld" to call "print_job_status_all".
The status line "[1] + Running exec 1>fifo" was being printed before the
input line "cat fifo" was read, causing the test case to fail.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3789 048f04df-13f5-43d7-8114-9f9ceecaec24