Substitute aliases anywhere in function definition
After alias substitution in which the alias value ends with a blank, the
next token is also subject to alias substitution. This POSIX requirement
applies anywhere including function definition commands. To support this
behavior (again), alias substitution function needs to be called from
many places in the function definition command parser, which is now
called after (rather than before) the simple command parser to eliminate
unneeded re-parsing and re-alias-substitution.

The is_literal_function_name function is introduced which decides if a
word is applicable for the function name used in a function definition
that does not begin with the "function" keyword. This replaces part of
the print_function_definition function, making the decision more correct
in terms of POSIXly-correctness and applicability of digits at the
beginning of the word.