Remove AF_NORECUR
Results of alias substitution are subject to another substitution, but
the second substitution should apply only after the resultant words have
been confirm not to be reserved words. That means the grammatical
context for the resultant words should be examined after every
substitution. It is not correct to substitute them in direct recursion
in a single invocation of substitute_alias.

To enforce indirect recursion of alias substitution, the AF_NORECUR
option is removed from the substitute_alias function.

Callers of substitute_alias does not yet implement correct recursion
considering grammatical context.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3803 048f04df-13f5-43d7-8114-9f9ceecaec24