Consider line continuations in all alias names
This commit is basically a refactoring to extract the common pattern in
calling the alias substitution function. However, it slightly modifies
the behavior of the parser and fixes a bug where line continuations are
not being parsed as expected in recursive alias substitution.

The count_name_length function, which removes line continuations in the
way, is now called not only before the first substitution but also
before every substitution.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3807 048f04df-13f5-43d7-8114-9f9ceecaec24