Open no-clobber redirection atomically
The previous implementation of the ">" redirection with the no-clobber
option was unreliable. After the is_irregular_file call, if another
process replaced the irregular file with a regular file, the shell was
clobbering the regular file.

To open the irregular file safely, the shell needs to open without the
O_CREAT or O_TRUNC flag and then check if it is really irregular.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3830 048f04df-13f5-43d7-8114-9f9ceecaec24