Use '\a' for fsl (#38240)
Background: In the default yashrc file, we use the tsl and fsl terminal
capabilities to modify the status line (or the window title) of the
terminal. Terminfo databases often lack those capabilities, so we have
been providing default values for common terminals depending on $TERM.
The default fsl for xterm had been '\033\\', which is defined as STRING
SEPARATOR in ECMA-48.

Problem: Some terminal emulators like Konsole do not support '\033\\' as
fsl and still pretend to be xterm with a false $TERM value. Our default
fsl results in a garbage shown in such terminals.

Solution: Many existing terminals support '\a' as fsl, which is now
considered a de facto standard. This should work on more terminals.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3838 048f04df-13f5-43d7-8114-9f9ceecaec24