Define prompt with $YASH_PSxy variables
The interpretation of the $PSxy variables differs depending on whether
the shell is in the POSIXly-correct mode or not. If the user switches
the mode on or off, the variables had to be redefined to match the new
interpretation. The shell now allows using the $YASH_PSxy variables
instead of $PSxy for defining prompt strings. As $YASH_PSxy are not used
in the POSIXly-correct mode, the user can define $PSxy to have different
values than $YASH_PSxy.

$PS1 is now always initialized to "$ " or "# ", as was only in the
POSIXly-correct mode. This default prompt works both in and out of the
POSIXly-correct mode.

Note: For backward compatibility with old initialization scripts which
only define $PSx to override the default prompt strings, we should not
initialize YASH_PSx by default.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3850 048f04df-13f5-43d7-8114-9f9ceecaec24