Test more complex alias substitution
In the "for" and "case" commands, the "in" and "do" keywords may appear
as a result of alias substitution that is initiated by another alias
substitution with a blank-ending value. Such "in" and "do" can be
preceded by newlines. This commit modifies existing tests to test that
newlines that appear from alias substitution are skipped before the
keywords are parsed.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3855 048f04df-13f5-43d7-8114-9f9ceecaec24