Count line number correctly after rewinding index (#38586)
This commit adds two test cases, the second of which is commented out as
I am not going to fix it soon. Since line continuations are entirely
removed from `ps->src' on the first pass, we cannot re-increment the
line number when incrementing the index over the removed line
continuation on the second pass.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3857 048f04df-13f5-43d7-8114-9f9ceecaec24