Always reject unclosed arithmetic expansion
According to POSIX, it is a syntax error if the shell reaches the end of
input before determining whether an expansion starting with "$((" is a
valid arithmetic expansion. Previously, yash treated such a case as if
it was determined not being an arithmetic expansion, then re-parsed it
as a command substitution. In most cases, the result was an invalid
command substitution, which caused a syntax error anyway. However, there
are corner cases where the expansion can be re-parsed as a valid command
substitution. Yash now correctly rejects it.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/branches/token_based_parser@3874 048f04df-13f5-43d7-8114-9f9ceecaec24