Neglect jobs in subshell with no-fork optimization (#38823)
When the no-fork optimization takes effect on a subshell, it was not
forgetting the original shell's jobs. The subshell was able to falsely
await the original shell's jobs.

The subshell should behave as if it is a child process created by the
fork_and_reset function. This commit extracts part of it as the
become_child function so that it can be directly called in the main
shell process performing the no-fork optimization.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3969 048f04df-13f5-43d7-8114-9f9ceecaec24