typeset: Print reparsable simple command starting w/ keyword (#37634)
The print_one_command function always prints redirections after the
other command components. As a result, a simple command printed could
not be re-parsed by the shell if the command name word was preceded by a
redirection (but not by any assignment) and the word would otherwise be
a keyword.

Such a word is now quoted so that it can successfully be re-parsed even
though it is printed as the first word of the command.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@3971 048f04df-13f5-43d7-8114-9f9ceecaec24