Optimize process forking with pipefail
When the pipe-fail option is on, the shell needs to await not only the
last command of a pipeline but also other commands. Because of this, the
optimization of not forking an unnecessary process for the last command
was not applied when the option was on.

If there is only one command in the pipeline, however, the pipe-fail
option can be ignored and the optimization is applicable.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@4009 048f04df-13f5-43d7-8114-9f9ceecaec24