Optimize read when stdin is a regular file
POSIX requires that the shell do not read too much from the standard
input. For this reason reading from the standard input is not buffered
by default. Without buffering, the "read" syscall is called for each
byte to be read, which may degrade the performance of the shell.

After this commit, when the standard input is a regular file, the input
is buffered to improve the performance. After reading, the file
descriptor is repositioned so that it works as if there has been no
buffering.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@4010 048f04df-13f5-43d7-8114-9f9ceecaec24