Fork inside exec_simple_command
Previously the exec_simple_command function required the finally_exit
argument to be true when executed for CT_EXTERNALPROGRAM. I don't think
this implicit precondition very intuitive because it effectively
nullifies the finally_exit parameter.

Starting from this commit we fork inside the exec_simple_command
function to allow the parameter to be false.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@4011 048f04df-13f5-43d7-8114-9f9ceecaec24