The new expand_four function
The beginning part of the expand_multiple and expand_word functions are
the same. This commit extracts the common part as the new expand_four
function and the former expand_four function is renamed to
expand_four_inner.

The expand_four_T structure is also restructured to match the behavior
of the new functions.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/branches/charcategory@4070 048f04df-13f5-43d7-8114-9f9ceecaec24