Faster brace expansion (#40708)
Previously, the time complexity of the brace expansion algorithm for an
input string of the form "{{{{{{...{{{" was O(n^2) where n is the length
of the input string. It was because, for each '{', the algorithm tries
to search all the remaining characters for a matching '}'.

In the new algorithm, we first scan the entire input string in O(n) time
and then generates the final results using the scan result. It should
work faster in case the expansion results in the same string as the
input.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@4087 048f04df-13f5-43d7-8114-9f9ceecaec24