configure: add -std=c99 conditionally (#45739)
Many users seem to try to override the default CC and/or CFLAGS without
knowing the effect of what they are doing. Our build configuration needs
a confirming C99 compiler, which may require a specific option to select
the language version. This commit adds a dynamic test that makes sure
the compiler is invoked with appropriate options.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@4209 048f04df-13f5-43d7-8114-9f9ceecaec24