Alias ending with blank & line continuation (#46532)
If the value of an alias substitution ends with a blank, the shell
checks the next token for another substitution. Previously, yash was
examining the parse buffer content to see if the last character of the
alias value was blank. However, this procedure incorrectly ignored any
line continuation that might be at the end of the value. After this
commit, the shell will test the last character in the alias definition
to determine the character blankness correctly.

git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/yash/yash/trunk@4233 048f04df-13f5-43d7-8114-9f9ceecaec24