Make run-test.sh POSIXly-portable
In this commit, the test runner script and test cases are modified in an
attempt to make the test runner runnable not only on yash but also on
any POSIX shell. Changes are also applied in this commit to work around
some bugs in existing shells. Below is a summary of the changes made in
this commit:

* The test runner script no longer aborts when run on a shell other than
  yash.
* The non-portable "-X" option for the "export" command is no longer
  used.
* Redirection error on the "exec" built-in aborts the POSIX shell, so file
  descriptors are now tested without a command.
* Test cases should not assume a particular process forking strategy for
  subshells, so the test for $PPID is rewritten.
* Some shells including dash and zsh would not like to have some shell
  variables unset. The $OPTARG variable is no longer unset.
* The "command -v" command does not print a correct absolute pathname to
  the given command on some shells including dash, hence the redundant
  "absolute" call.
* Some shells print an error message when a process is killed by a
  signal, which was causing some test cases to fail unexpectedly.
  Redirections on the testee are now applied in a subshell so that the
  message does not affect the test result.
* Ksh seems to be tricked by file descriptors applied for other
  commands. To avoid false failures, unused file descriptors are
  explicitly closed when calling the "testcase" function.

Unfortunately, at the time of this commit, only Bash is known to run all
the tests successfully other than yash. Dash, ksh, mksh, and zsh each
have unresolved issues that prevent the tests from being run
successfully. It is unknown whether those issues are due to bugs in the
shells or in the tests.