Reorganize directory structure
GitHub pages does not support the language selecting extension, so we
now have separate directories for different languages.

The makenav script is removed as it was never used.

The link to the wiki is removed from the Japanese index.html as it is
not very likely that the wiki will retain many Japanese contents.