tests/kill4: Fix tests with wrong assumption
Previously, some tests in kill4-p.tst depended on traps reacting to a
signal sent by the shell. Since the traps were set in subshells, the
processes executing the trap built-in might not be the direct child
processes of the shell. In that case, the children were terminated by
the signal, possibly before the subshells responded to the signal,
causing the subsequent wait built-in to return too early in the main
shell process.

To resolve this issue, the test cases are rewritten without using traps.