tests/wait-p: Don't depend on yash-specific behavior
Previously, the test case 'trap interrupts wait' assumed that the
asynchronous subshell ran as a direct child process of the main shell.
This assumption was not POSIXly-portable. The shell is allowed to run
the subshell in a child process of the asynchronous job's process, in
which case the job process might be terminated by SIGUSR2 before the
trap action is performed in the subshell.

This commit rewrites the test so that the test result does not depend on
the wait status for the subshell.