Account for existing backslashes in completion
The previous implementation assumed that there was no backslash escape
preceding the cursor when performing command line completion. This
caused a redundant backslash to be inserted when the next character in
the completed word needed escaping. After this commit, the shell checks
for an escape before the current position and removes it when inserting
the completion, leaving the correct number of backslashes.

Fixes https://github.com/magicant/yash/issues/47