Skip tests of test if the shell does not implement it
If the testee does not implement the test utility as a built-in, the
test cases in tests/test-p.tst tests the behavior of the external
utility found in $PATH, which may (or may not) affect the results.
We should skip the tests if we are not sure if the shell supports the
built-in.