Reorganize tests for quotes in parameter expansion
In POSIX.1-2024, it is clarified that the word in ${param=word} is
subject to quote removal, so the relevant behavior now can be tests in
quote-p.tst instead of quote-y.tst. It is also clarified that } can be
escaped with a backslash in parameter expansions, which allows moving
the relevant tests as well.