Position specifier (%n$) in printf built-in
This commit adds support for position specifiers in the printf built-in
as required by POSIX. A position specifier is a dollar sign ($) preceded
by a positive decimal integer. The integer specifies the index of the
value to be formatted. For example, in the format string '%2$s %1$s',
the second value is formatted first and the first value is formatted
next.

Closes https://github.com/magicant/yash/issues/102