trap: Implement POSIX.1-2024 requirements
This commit implements the requirements of POSIX.1-2024 for the `trap`
built-in. The `trap` built-in now supports the `-p` option in the
POSIXly-correct mode. When the `-p` option is used, the built-in prints
all traps without filtering out traps that are not set. The built-in
also reports signals that are ignored because the program that invoked
the shell has set the signal to be ignored.

The test case file is synced with:
https://github.com/magicant/yash-rs/blob/c348c1f0fed18be5b760bd1ed5ec73f224e53731/yash-cli/tests/scripted_test/trap-p.sh