history: Work around glibc's fsetpos bug
The fsetpos function in glibc seems to be buggy. It sometimes sets the
FILE's error indicator without actually setting errno. This causes the
shell to fail to update the history file. To work around this issue, the
shell now checks the error indicator after calling fsetpos and re-reads
the history file from the beginning if it is set.

The check script for working I/O functions in configure is updated in an
attempt to reproduce the issue, but currently it does not catch the bug
(at least on my environment with glibc 2.40 on Fedora 41).