wait: Don't return when the target process is suspended
The `wait` built-in no longer returns when the target process is
suspended if job-control is on. The built-in now waits for the target
process to exit or be terminated by a signal regardless of mode.
This behavior is required by POSIX.1-2024.

Maybe we could have a new option to make the built-in work as before,
but I'd like to see if other shells introduce such an option first.