Ignore remaining commands after a foreground job stops
POSIX.1-2024 requires that when a foreground job stops in an
interactive, job-control shell, the shell discards any remaining
commands that were typed but not yet executed, up to and including a
next asynchronous list. This is to prevent the user from accidentally
executing commands that were intended to be executed after previous
commands have completed.

Yash did not previously implement this behavior. This commit implements
the behavior, but not exactly as specified in POSIX.1-2024. Instead of
discarding commands up to and including the next asynchronous list, yash
discards all commands that were read but not executed. This is to keep
the implementation simple and should be more intuitive for users. See
<https://github.com/magicant/yash-rs/issues/478> for more discussion.

To implement this behavior, the shell raises an exception when a
foreground job stops. The exception is cleared when the shell reads a
new command.