Improve test coverage for behavior after a job is suspended
This commit adds another invocation of `fg`, in which the job is
suspended, to the test case "discard remaining commands when a command
suspends". Previously, the test case only covered the behavior when an
initially foreground job is suspended. This commit adds coverage for the
behavior when a resumed job is suspended.