Treat a job as suspended when any child process suspends
POSIX.1-2024 requires that the shell treat a job as suspended when any
child process comprising the job suspends. Previously, the shell
considered a job to be suspended only when all child processes
comprising the job suspended. This change makes the shell compliant
with the standard.

Fixes https://github.com/magicant/yash/issues/130