Open ttyfd with O_NOCTTY
POSIX.1-2024 Job control specifications are written in the assumption
that a job-control shell may not have a control terminal. The shell
should not make an arbitrary terminal its control terminal. To make sure
that the shell only opens an existing control terminal, this commit adds
O_NOCTTY to the open(2) call for /dev/tty.