Allow job control without a controlling terminal
The wording of POSIX.1-2024 implies that a shell can perform job control
even if it does not have a controlling terminal. This commit allows yash
to enable job control even if it does not have a controlling terminal.

To do this, the `is_interactive_now` variable is removed, and a new
`is_subshell` variable is added. The `doing_job_control_now` macro is
modified to check if the shell is a subshell. The `open_ttyfd` function
is modified not to disable job control if it fails to open the terminal
file.

As a side effect, the `-m` option can no longer be used to enable job
control in a subshell. This behavior was not documented and probably
never used.